// src/api/generated/models/Cigar.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Cigar = {
    id: string;
    humidor_id: string;
    name: string;
    brand?: string | null;
    vitola?: string | null;
    year?: number | null;
    notes?: string | null;
};